run clippy and fmt

+++ kernel/arch/x64/interrupts/handlers.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::types::PowerOfTwoOps;
-
 use super::exceptions::ExceptionFrame;
+use crate::types::PowerOfTwoOps;
 
 extern "C" {
     fn stack_guard_top();
@@ -17,9 +16,9 @@ pub(super) fn divide_by_zero(_frame: &ExceptionFrame) {
 
 pub(super) fn page_fault(_frame: &ExceptionFrame) {
     let vaddr = read_reg!(cr2);
-    let round = vaddr.page_round_down();
-    let guard_top_low = (stack_guard_top as u64) & 0xffffffff;
-    let guard_bot_low = (stack_guard_bot as u64) & 0xffffffff;
+    let round = vaddr.page_round_down() as usize;
+    let guard_top_low = (stack_guard_top as usize) & 0xffffffff;
+    let guard_bot_low = (stack_guard_bot as usize) & 0xffffffff;
 
     if round == guard_top_low {
         panic!("Kernel stack overflow");
